unit test

+++ core/src/main/java/com/sstudio/submissionbajetpackpro/core/domain/model/MovieDetail.kt
@@ -3,7 +3,7 @@ package com.sstudio.submissionbajetpackpro.core.domain.model
 data class MovieDetail(
     val movie: Movie,
     val adult: Boolean? = null,
-    val budget: Int = 0,
+    val budget: Long = 0,
     val genres: List<Genre> = listOf(),
     val homepage: String = "",
     val id: Int = 0,
@@ -11,7 +11,7 @@ data class MovieDetail(
     val originalLanguage: String = "",
     val popularity: Double = 0.0,
     val productionCompanies: List<ProductionCompany> = listOf(),
-    val revenue: Int = 0,
+    val revenue: Long = 0,
     val runtime: Int = 0,
     val status: String = "",
     val title: String = "",
